requirements.txt
apache-airflow==2.7.3
apache-airflow-providers-snowflake==5.1.0
pyspark==3.4.1
pandas==2.1.0
pyarrow==13.0.0
boto3==1.28.0
snowflake-connector-python==3.5.0
snowflake-sqlalchemy==1.5.0
great-expectations==0.18.0
pyyaml==6.0.1
